--- src/domain/mod.rs
@@ -1,3 +1,0 @@
-mod new_image;
-
-pub use new_image::NewImage;
